Fix Prawn::QRCode#print_qr_code
This calls the `Renderer` with a `margin` argument of `margin`. However,
`#margin` is not defined on the context of this instance method.

If I understand it correctly, this should pass on the `margin` option
from the options hash (same as with the `stroke` argument). This is
accomplished by the `**options` double splat at the end of the method
call.